fix empty state message in location bookings table
